+++ src/java/models/ModeloOrders.java
@@ -28,7 +28,7 @@ public class ModeloOrders extends Conexion {
             pst = getConnection().prepareCall(sql);
             rs = pst.executeQuery();
             while (rs.next()) {
-                ordenes.add(new Orders(rs.getInt("OrderID"), rs.getInt("OrderUserID"), rs.getFloat("OrderAmount"), rs.getString("OrderShipName"), rs.getString("OrderShipAddress"), rs.getString("OrderPhone"), rs.getFloat("OrderShipping"), rs.getString("OrderEmail"), rs.getString("OrderDate"), rs.getBoolean("OrderShipped"), rs.getString("OrderTrackingNumber")));
+                ordenes.add(new Orders(rs.getInt("OrderID"), rs.getInt("OrderUserID"), rs.getInt("OrderAmount"), rs.getString("OrderShipName"), rs.getString("OrderShipAddress"), rs.getString("OrderPhone"), rs.getInt("OrderShipping"), rs.getString("OrderEmail"), rs.getString("OrderDate"), rs.getBoolean("OrderShipped"), rs.getString("OrderTrackingNumber")));
             }
 
         } catch (Exception e) {
@@ -61,7 +61,7 @@ public class ModeloOrders extends Conexion {
             pst.setInt(1, id);
             rs = pst.executeQuery();
             while (rs.next()) {
-                orden = new Orders(rs.getInt("OrderID"), rs.getInt("OrderUserID"), rs.getFloat("OrderAmount"), rs.getString("OrderShipName"), rs.getString("OrderShipAddress"), rs.getString("OrderPhone"), rs.getFloat("OrderShipping"), rs.getString("OrderEmail"), rs.getString("OrderDate"), rs.getBoolean("OrderShipped"), rs.getString("OrderTrackingNumber"));
+                orden = new Orders(rs.getInt("OrderID"), rs.getInt("OrderUserID"), rs.getInt("OrderAmount"), rs.getString("OrderShipName"), rs.getString("OrderShipAddress"), rs.getString("OrderPhone"), rs.getInt("OrderShipping"), rs.getString("OrderEmail"), rs.getString("OrderDate"), rs.getBoolean("OrderShipped"), rs.getString("OrderTrackingNumber"));
             }
 
         } catch (Exception e) {
@@ -173,7 +173,7 @@ public class ModeloOrders extends Conexion {
             String sql = "UPDATE orders SET OrderAmount=?,OrderShipAddress=?,OrderPhone=?,OrderEmail=?, OrderShipped=? WHERE OrderID = ?";
             pst = getConnection().prepareStatement(sql);
             
-            pst.setFloat(1, orden.getOrderAmount());
+            pst.setInt(1, orden.getOrderAmount());
             pst.setString(2, orden.getOrderShipAddress());
             
             pst.setString(3, orden.getOrderPhone());
